Working LHE and Gen weights
