+++ python/nano_cff.py
@@ -47,12 +47,14 @@ genWeightsTable = cms.EDProducer("GenWeightsTableProducer",
     name = cms.string("genEvent"),
     doc  = cms.string("generator weights from GenEventInfoProduct (main per-event weight, may be negative)"),
     genEvent = cms.InputTag("generator"),
-)
-lheInfoTable = cms.EDProducer("LHETablesProducer",
     lheInfo = cms.InputTag("externalLHEProducer"),
     preferredPDFs = cms.vuint32(91400,260001),
     namedWeightIDs = cms.vstring(),
     namedWeightLabels = cms.vstring(),
+    debug = cms.untracked.bool(False),
+)
+lheInfoTable = cms.EDProducer("LHETablesProducer",
+    lheInfo = cms.InputTag("externalLHEProducer"),
 )
